Moved ScheduleMedalRound files to Schedule2016/MedalRound directory.
No functionality changes.  I hope.  IDE does flag some stuff but I did not go through and "clean" it up.
